requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
boto3>=1.29.0
pydantic>=2.5.0
python-multipart>=0.0.6
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
httpx>=0.25.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
aiohttp>=3.9.0
tenacity>=8.2.3